# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "在线聊天系统",
        "title": "chat",
        "contact": {},
        "license": {
            "name": "lyc,why"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8084",
    "basePath": "/",
    "paths": {
        "/api/account/info": {
            "get": {
                "consumes": [
                    "application/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "获取账户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "account_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2009:权限不足 2007:身份不存在 2008:身份验证失败 2010:账号不存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.GetAccountByID"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/account/infos/name": {
            "get": {
                "consumes": [
                    "application/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "通过昵称模糊查找账户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "Page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "PageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "account_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2010:账号不存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.GetAccountsByName"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/account/infos/user": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "创建账户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1003:系统错误 2008:身份验证失败 2010:账号不存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.GetAccountsByUserID"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/account/update": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "更新账户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-token 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "账号信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateAccount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
...
        "/api/v1/account//infos/user": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "获取用户的所有账号",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 ",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.TotalAccountsReply"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/account/deleteAccount/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "删除账户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "删除账号的ID",
                        "name": "accountID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 ",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/v1/account/getToken/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "获取账户令牌",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "删除账号的ID",
                        "name": "accountID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 ",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/common.Token"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.State": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "失败时返回空"
                },
                "status_code": {
                    "description": "状态码，0-成功，其他值-失败",
                    "type": "integer"
                },
                "status_msg": {
                    "description": "返回状态描述",
                    "type": "string"
                }
            }
        },
        "common.Token": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "reply.AccountInfo": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "reply.AccountInfoReply": {
            "type": "object",
            "properties": {
                "accountID": {
                    "type": "integer"
                },
                "avatar": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
...
        "reply.GetAccountByID": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "reply.GetAccountsByName": {
            "type": "object",
            "properties": {
                "accountInfos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reply.AccountInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "reply.GetAccountsByUserID": {
            "type": "object",
            "properties": {
                "accountInfos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reply.AccountInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "reply.TotalAccountsReply": {
            "type": "object",
            "properties": {
                "accountInfos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reply.AccountInfoReply"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "request.UpdateAccount": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}